feat: finish modularization of MediaWiki API MCP client
We are currently implementing an MCP server that allows LLMs like Claude to interact with a MediaWiki installations, as a bot user, through the MediaWiki API interface.

We needed to do some refactoring of our MediaWiki API MCP but we were interrupted. The `client.py` file was getting too long, which has impact on maintainability and other concerns. Therefore we need to modularize it. We decided to keep the `client.py` file as the orchestrating file (on the same level as `server.py`), and to put the modules in a `client_modules` directory. Additionally, the module files are named following the format `client_page.py`, `client_search.py`, etc. Finally, pay specific attention should be paid to the `MediaWikiConfig` class which was originally (and still is?) located in the `client.py` file. Please finish the job.

```git-revs
2e998b7  (Base revision)
154ef02  Create config.py with MediaWikiConfig class
0c286be  Update client.py to import MediaWikiConfig from config module
dc95b9b  Update server.py to import MediaWikiConfig from config module
96419b4  Update wiki_page_edit.py imports to use config module
ba4893d  Update wiki_page_get.py imports to use config module
1dd0f5a  Update wiki_search.py imports to use config module
bda9e3a  Update wiki_opensearch.py imports to use config module
HEAD     Create script to update remaining server_tools imports
```

codemcp-id: 35-feat-finish-modularization-of-mediawiki-api-mcp-cl
