+++ mediawiki_api_mcp/config.py
@@ -1,0 +1,11 @@
+"""Configuration for MediaWiki API MCP integration."""
+
+from pydantic import BaseModel
+
+
+class MediaWikiConfig(BaseModel):
+    """Configuration for MediaWiki API connection."""
+    api_url: str
+    username: str
+    password: str
+    user_agent: str = "MediaWiki-MCP-Bot/1.0"
